// src/app/model/pots-task.ts
export class PotsTask {
    title: string;
    description: string;
    domainID: string;
    departmentID: number;
    categoryID: number;
    ownerID: number;
    secondaryOwnerID: number;
    taskStatus: number;
    performerID: number;
    secondaryPerformerID: number;
    reviewerID: number;
    approverID: number;
    frequencyID: number;
    startDate: number;
    endDate: number;
    remindDays: number;
}